+++ authority-portal-frontend/src/app/core/utils/array-utils.ts
@@ -28,3 +28,8 @@ export function removeOnce<T>(list: T[], item: T): T[] {
   }
   return list;
 }
+
+export function filterNonNull<T>(array: (T | null | undefined)[]): T[] {
+  return array.filter((it) => it != null) as T[];
+}
+